Adding a blog post content
This creates a markdown file example of a blog post, and related image files.

Ideally the images will be responsive using srcset so each image will need to be created at multiple sizes. This will need a further spike to work out the best way to handle responsive images with native lazy loading, but for now single images will be enough for a first pass.