Responsive layout
Refactored css to style the rendering of a post on mobile.

The pullquote style does not make sense on a narrow layout and is removed from
the smaller viewport width.

Separated out styling that is related to the 12 column grid and general styling
for the post itself. This allows flexibility in future to change the grid without
having to repeat shared styling for headings etc.