CSS refinements
This removes `panel` utility class from the post `time` element to instead
utilise the `flow` utility class to control the spacing between the date and heading.

Image row span extended by an extra row to ensure that the date and title butt up
underneath eachother as expected, rather than spacing out to match the height of
the featured image.