Extending markdown
2 plugins have been added to extend markdown capabilities to make blog post writing
easier in markdown.

ids, classes and aria attributes can now be added in markdown by appending to the
end of a block eg.
```
a paragraph with customattributes{#id .custom-class aria-hidden=true}
```

The plugin recommends creating a whitelist for security. Although it shouldn't be
an issue, I've created one to only allow ids, classes and attributes staring with `aria`.

can create description (definition) lists in markdown
```
This will be the dt text
: this will be the dd text
```