Refactor collections partial
The collections partial does not currently take into account the latest post and
there is no method to set the number of posts displayed in a collection.

This adds `getSiblingContent` which will take a passed item, usually the latest post,
and exclude it from the required collection. It will also limit the number of items
returned if the limit flag is passed. By default a limit is not applied.

The `blog` page implementation of the `collections` partial is updated to exclude
the latest post and limit total number of posts displayed under the three main tags
to a max of 5.